refactor(client): simplify — collapse the abort forwarder + extract a per-host prefs owner member
/be-review simplify pass (3 findings; reuse + efficiency were clean):
- Collapse beginAuthoritativeMutation (now a pure pass-through to the codex-extracted
  abortTransientInput) into abortTransientInput at its 3 call sites; fold its doc note.
- ALTITUDE: the sticky per-host prefs (activityWindow/showSleeping/rightPanelCollapsed)
  were lumped into createViewState, forcing reset() to hand-enumerate which signals to
  clear vs keep — the 'forgotten field' enumeration hazard W7 exists to kill. Extract
  createHostPrefs (a 4th HostScope owner member: {view, prefs, camera, restore});
  createViewState now holds only reset-on-close-all selection+posture and reset() clears
  its whole factory unconditionally. Facades re-point activeScope()?.view.X -> .prefs.X.
- Encapsulate the per-host storage-key suffix in a perHostKey closure inside createHostPrefs
  so a per-host pref can't be spelled without being host-scoped. Exact keys preserved.

635 client tests + typecheck + just fmt green; perHostCanvas.test's 3 it() pins byte-identical.
