+++ packages/client/src/hostScope/createViewState.ts
@@ -10,27 +10,24 @@
  *  them across a switch-away is free); they survive a switch-away in this owner
  *  and are disposed only when the host leaves `padiMap.entries`.
  *
- *  The camera moved OUT to a sibling owner member (`createCamera`) — it is no
- *  longer a `HostView` field. W7 TIER A pulled MORE per-host view facts in here:
- *  the fullscreen posture (`canvasMaximized`), the two dock filters
- *  (`activityWindow`, `showSleeping`), and the right-panel collapsed bit — each a
- *  VIEW OF this host's content, so each is per-host by THE RULE (see
- *  `canvas/canvasBoundaryGuard.test.ts`). Only the momentary `centerActiveRequest`
- *  command stays APP-level in the facade — a write-and-consume viewport impulse,
- *  never durable per-host state. The dock filters persist PER HOST in localStorage
- *  (a sticky preference must survive reload); the postures are in-memory (reload
- *  → tiled/inherit, matching the camera tier). */
+ *  The camera and the sticky per-host PREFERENCES moved OUT to sibling owner
+ *  members (`createCamera`, `createHostPrefs`) — they are no longer `HostView`
+ *  fields. What W7 TIER A DID leave here is the fullscreen posture
+ *  (`canvasMaximized`) — a VIEW OF this host's content (per-host by THE RULE, see
+ *  `canvas/canvasBoundaryGuard.test.ts`) that, like the selection facts, a
+ *  close-all `reset()` clears. That split is deliberate: everything left in this
+ *  factory is reset-on-close-all, so `reset()` clears its WHOLE state with no
+ *  "clear these but not those" allow/deny list (the enumeration hazard W7 kills).
+ *  The sticky filters/collapsed bit live in `createHostPrefs` precisely because
+ *  they must SURVIVE a close-all. Only the momentary `centerActiveRequest` command
+ *  stays APP-level in the facade — a write-and-consume viewport impulse, never
+ *  durable per-host state. The posture is in-memory (reload → tiled, matching the
+ *  camera tier). */
 
 import { encodeHostKey, type HostKey } from "kolu-common/hostKey";
 import type { TerminalId } from "kolu-common/surface";
 import { type Accessor, createSignal, type Setter } from "solid-js";
 import { createStore, produce, reconcile } from "solid-js/store";
-import { boolPref, persistedPref } from "../persistedPref";
-import {
-  type ActivityWindow,
-  DEFAULT_ACTIVITY_WINDOW,
-  isActivityWindow,
-} from "../terminal/activityWindow";
 import { padiRpcOf } from "../wire";
 
 type TerminalAttention = "unread" | "badge-only";
@@ -50,27 +47,14 @@ export interface HostViewState {
   clearBadgeAttention: () => void;
   isUnread: (id: TerminalId) => boolean;
   hasBadgeAttention: (id: TerminalId) => boolean;
-  // ── Per-host VIEW POSTURE + dock filters (W7 TIER A) ─────────────────
+  // ── Per-host VIEW POSTURE (W7 TIER A) ────────────────────────────────
   /** Fullscreen-one-tile posture for THIS host. In-memory (reload resets to
    *  tiled — the camera tier), per-host so a switch shows each host's own
-   *  posture. */
+   *  posture. Cleared by `reset()` (a close-all drops back to tiled). The sticky
+   *  dock filters + right-panel bit live in `createHostPrefs` — they SURVIVE a
+   *  close-all, so they are NOT here. */
   canvasMaximized: Accessor<boolean>;
   setCanvasMaximized: Setter<boolean>;
-  /** This host's dock activity-window filter — persisted per host under
-   *  `kolu-activityWindow:<encoded host>` so a host's filter survives reload (a
-   *  sticky dock preference, unlike the volatile camera/posture) without two
-   *  hosts colliding on one global key. */
-  activityWindow: Accessor<ActivityWindow>;
-  setActivityWindow: Setter<ActivityWindow>;
-  /** Whether THIS host's dock shows sleeping (☾) rows — persisted per host under
-   *  `kolu-showSleeping:<encoded host>`, same rationale as `activityWindow`. */
-  showSleeping: Accessor<boolean>;
-  setShowSleeping: Setter<boolean>;
-  /** This host's right-panel collapsed bit — `undefined` means "inherit the
-   *  global preference" (the seed the facade reads on first sight); a set value
-   *  is this host's in-memory override (reload re-inherits the global). */
-  rightPanelCollapsed: Accessor<boolean | undefined>;
-  setRightPanelCollapsed: Setter<boolean | undefined>;
   reset: () => void;
 }
 
@@ -81,35 +65,12 @@ export function createViewState(host: HostKey): HostViewState {
     Record<TerminalId, TerminalAttention>
   >({});
 
-  // The canonical host string — the per-host storage-key suffix + the map's
-  // `codec.encode(host)`. Computed once per owner.
+  // The canonical host string — the map's `codec.encode(host)`, used in the
+  // active-terminal report's error message below. Computed once per owner.
   const encoded = encodeHostKey(host);
 
   // View posture: in-memory per host (reload → tiled, matching the camera tier).
   const [canvasMaximized, setCanvasMaximized] = createSignal(false);
-
-  // Dock filters: persisted PER HOST — a dock filter is a sticky preference (it
-  // must survive reload), but keyed by host so two hosts don't share one filter.
-  const [activityWindow, setActivityWindow] = persistedPref<ActivityWindow>({
-    name: `kolu-activityWindow:${encoded}`,
-    fallback: DEFAULT_ACTIVITY_WINDOW,
-    parse: (raw) => {
-      if (isActivityWindow(raw)) return raw;
-      throw new Error(`unrecognized activity window: ${raw}`);
-    },
-  });
-  const [showSleeping, setShowSleeping] = boolPref({
-    name: `kolu-showSleeping:${encoded}`,
-    fallback: true,
-  });
-
-  // Right-panel collapsed: in-memory per host. `undefined` = inherit the global
-  // preference — the `useRightPanel` facade seeds the read with
-  // `preferences().rightPanel.collapsed`, so a host's first sight matches the old
-  // global bit and only diverges once the user toggles it on THAT host.
-  const [rightPanelCollapsed, setRightPanelCollapsed] = createSignal<
-    boolean | undefined
-  >(undefined);
 
   function writeActive(id: TerminalId | null): void {
     setActiveId(id);
@@ -171,13 +132,15 @@ export function createViewState(host: HostKey): HostViewState {
   }
 
   function reset(): void {
+    // This factory now owns ONLY reset-on-close-all state — the selection facts
+    // (active tile, MRU, attention) plus the maximized posture. The sticky per-host
+    // PREFERENCES moved to `createHostPrefs`, so there is no "clear these but not
+    // the prefs" allow/deny list to keep in sync: `reset` unconditionally clears
+    // every signal this factory owns. Closing every tile drops this host back to
+    // the tiled posture (matching the pre-per-host `reset` clearing `canvasMaximized`).
     setActiveId(null);
     setMru([]);
     setAttention(reconcile({}));
-    // Closing every tile drops this host back to the tiled posture (matching the
-    // pre-per-host behavior where `reset` cleared `canvasMaximized`). The dock
-    // filters and the right-panel bit are sticky preferences, NOT selection state,
-    // so a close-all leaves them untouched.
     setCanvasMaximized(false);
   }
 
@@ -193,12 +156,6 @@ export function createViewState(host: HostKey): HostViewState {
     hasBadgeAttention,
     canvasMaximized,
     setCanvasMaximized,
-    activityWindow,
-    setActivityWindow,
-    showSleeping,
-    setShowSleeping,
-    rightPanelCollapsed,
-    setRightPanelCollapsed,
     reset,
   };
 }
