+++ packages/client/src/perHostCanvas.test.ts
@@ -98,7 +98,7 @@ vi.mock("./wire", async () => {
   };
 });
 
-// `createViewState`'s per-host prefs (`showSleeping` via `boolPref`, `activityWindow`
+// `createHostPrefs`'s per-host prefs (`showSleeping` via `boolPref`, `activityWindow`
 // via `persistedPref`) — stub both to a plain in-memory signal-shaped pair honoring
 // the passed `fallback`, so the test needs no real `localStorage`. The three blocks
 // below exercise only `activeId`/`mruOrder`, so these prefs' values are never asserted.
