+++ packages/client/src/terminal/activityWindowFilter.ts
@@ -4,7 +4,7 @@
  *  Kept OUT of the pure vocabulary leaf (`activityWindow.ts`, the `ActivityWindow`
  *  type + `WINDOWS` ladder + `isActivityWindow`/`windowOption`/idle-bucketing) so
  *  that least-volatile leaf has NO back-edge into the most-volatile per-host owner:
- *  `createViewState` imports only the vocab leaf, this facade imports the vocab leaf
+ *  `createHostPrefs` imports only the vocab leaf, this facade imports the vocab leaf
  *  plus `activeScope` — layering stays downward, no import cycle. */
 
 import { activeScope } from "../hostScope/hostScopes";
@@ -15,19 +15,19 @@ import {
 } from "./activityWindow";
 
 /** The ACTIVE host's activity-window choice — a per-host fact born in the host
- *  scope's `createViewState` (persisted per host under `kolu-activityWindow:<host>`),
+ *  scope's `createHostPrefs` (persisted per host under `kolu-activityWindow:<host>`),
  *  read here through the facade. W7 TIER A moved this OUT of one global localStorage
  *  singleton: switching hosts now shows each host's own dock filter, and the choice
  *  parameterizes a VIEW OF that host's terminals (per-host by THE RULE). Floors the
  *  removal race to the default, exactly as `useViewState` floors its per-host reads. */
 export function activityWindow(): ActivityWindow {
-  return activeScope()?.view.activityWindow() ?? DEFAULT_ACTIVITY_WINDOW;
+  return activeScope()?.prefs.activityWindow() ?? DEFAULT_ACTIVITY_WINDOW;
 }
 
 /** Set the ACTIVE host's activity window (a no-op during the one-tick removal
  *  race, when there is no active scope to write). */
 export function setActivityWindow(next: ActivityWindow): void {
-  activeScope()?.view.setActivityWindow(next);
+  activeScope()?.prefs.setActivityWindow(next);
 }
 
 /** Reactive threshold (ms) for the currently-selected activity window.
